refactor(process view): create LogoutButton component
refs #1294
